requirements.txt
pydantic>=2.8
requests>=2.32
pandas>=2.2
nltk>=3.9
vaderSentiment>=3.3.2
psycopg[binary,pool]>=3.2
SQLAlchemy>=2.0
python-dotenv>=1.0
loguru>=0.7
pyarrow>=16.1
